# Allow docs build without examples (much faster)