--- a/.readthedocs.yml
+++ b/.readthedocs.yml
@@ -16,7 +16,7 @@ formats:
 
 # Optionally set the version of Python and requirements required to build your docs
 python:
-  version: 3.11
+  version: 3.8
   install:
     - requirements: requirements_dev.txt
     - requirements: requirements.txt