--- a/.readthedocs.yml
+++ b/.readthedocs.yml
@@ -1,22 +1,30 @@
-
-# .readthedocs.yml
+# .readthedocs.yaml
 # Read the Docs configuration file
 # See https://docs.readthedocs.io/en/stable/config-file/v2.html for details
 
 # Required
 version: 2
 
+# Set the version of Python and other tools you might need
+build:
+  os: ubuntu-22.04
+  tools:
+    python: "3.11"
+    # You can also specify other tool versions:
+    # nodejs: "19"
+    # rust: "1.64"
+    # golang: "1.19"
+
 # Build documentation in the docs/ directory with Sphinx
 sphinx:
-  configuration: docs/conf.py
+   configuration: docs/conf.py
 
-# Doc formats that will be built. Can add pdf, epub if wanted. 
+# If using Sphinx, optionally build your docs in additional formats such as PDF
 formats:
-  - htmlzip
+   - htmlzip
 
-# Optionally set the version of Python and requirements required to build your docs
+# Optionally declare the Python requirements required to build your docs
 python:
-  version: 3.8
   install:
     - requirements: requirements_dev.txt
     - requirements: requirements.txt