--- a/docs/conf.py
+++ b/docs/conf.py
@@ -87,7 +87,7 @@ master_doc = "index"
 
 # General information about the project.
 project = "pint"
-copyright = "2017-2021 PINT Developers"
+copyright = "2017-2023 PINT Developers"
 
 # The version info for the project you're documenting, acts as replacement
 # for |version| and |release|, also used in various other places throughout
@@ -203,7 +203,7 @@ if not on_rtd:  # only import and set the theme if we're building docs locally
 
 # The name of an image file (relative to this directory) to place at the
 # top of the sidebar.
-# html_logo = None
+html_logo = "logo/PINT_LOGO_64trans.png"
 
 # The name of an image file (within the static path) to use as favicon
 # of the docs.  This file should be a Windows icon file (.ico) being