--- a/requirements_dev.txt
+++ b/requirements_dev.txt
@@ -6,7 +6,7 @@ traitlets
 Sphinx>=2.2,!=5.1.0
 astropy>=4.0,!=4.0.1,!=4.0.1.post1
 #astropy-helpers>=1.3
-sphinx-rtd-theme>=0.1.9
+sphinx-rtd-theme>=1.1.1
 coveralls>=1.1
 wheel>=0.29.0
 pytest>=4.3