Standard RTD builds take a long time.  About 6 min on my system.  Much of that time is spent running the example notebooks.  This gives a new build target for the docs that excludes them.  It takes about 90s on my system:
```
(pintdev) kaplan@plock[~/pythonpackages/PINT_dlakaplan/docs] (docsnoexamples) % make cleaner; time make html >& ../html.out
rm -rf _build/*
rm -rf _autosummary/*.rst
make html >&../html.out  367.42s user 15.96s system 96% cpu 6:35.46 total
```
vs.
```
(pintdev) kaplan@plock[~/pythonpackages/PINT_dlakaplan/docs] (docsnoexamples) % make cleaner; time make htmlnoexamples >& ../htmlnoexamples.out
rm -rf _build/*
rm -rf _autosummary/*.rst
make htmlnoexamples >&../htmlnoexamples.out  78.49s user 10.44s system 89% cpu 1:39.79 total
```
So you would run `make htmlnoexamples`

I also added `cleaner` which removes not only the files in `_build/` (done by standard `make clean`) but `_autosummary/`, since those may need to be regenerated when modules are added/removed.  Note that this directory is hard-coded since I couldn't find that it was set at the top level.

Finally, I updated the version of `sphinx-rtd-theme` so that search seems to work again,  and added the logo (back?) to the RTD page.