@aarchiba 2023-02-02
What does the documentation look like with this option? Does the tutorials section just show as blank? Dangling links? It would be nice to put in a message so we know what happened if we end up with a version generated with this flag.